+++ src/EcommerceApp.Application/Features/UserFeatures/Commands/DeleteUserCommand.cs
@@ -35,7 +35,7 @@ namespace Application.Features.UserFeatures.Commands
                 };
             }
 
-            if (request.Hard.HasValue)
+            if (request.Hard.HasValue && request.Hard.Value)
             {
                 await _unitOfWork.UserRepository.Delete(user);
             } else
